+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added a glossary hint icon to the Incomplete Argumentation acceptance type selector.
+
 ## [0.12.0] - 2026-09-21
 
 ### Added